+++ sdk/dependencymap/azure-resourcemanager-dependencymap/src/main/java/com/azure/resourcemanager/dependencymap/implementation/MapsImpl.java
@@ -10,10 +10,15 @@ import com.azure.core.http.rest.SimpleResponse;
 import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.dependencymap.fluent.MapsClient;
+import com.azure.resourcemanager.dependencymap.fluent.models.ExportDependenciesOperationResultInner;
+import com.azure.resourcemanager.dependencymap.fluent.models.GetDependencyViewForAllMachinesOperationResultInner;
 import com.azure.resourcemanager.dependencymap.fluent.models.MapsResourceInner;
+import com.azure.resourcemanager.dependencymap.models.ExportDependenciesOperationResult;
 import com.azure.resourcemanager.dependencymap.models.ExportDependenciesRequest;
 import com.azure.resourcemanager.dependencymap.models.GetConnectionsForProcessOnFocusedMachineRequest;
 import com.azure.resourcemanager.dependencymap.models.GetConnectionsWithConnectedMachineForFocusedMachineRequest;
+import com.azure.resourcemanager.dependencymap.models.GetDependencyViewForAllMachinesOperationResult;
+import com.azure.resourcemanager.dependencymap.models.GetDependencyViewForAllMachinesRequest;
 import com.azure.resourcemanager.dependencymap.models.GetDependencyViewForFocusedMachineRequest;
 import com.azure.resourcemanager.dependencymap.models.Maps;
 import com.azure.resourcemanager.dependencymap.models.MapsResource;
@@ -111,13 +116,48 @@ public final class MapsImpl implements Maps {
         this.serviceClient().getConnectionsForProcessOnFocusedMachine(resourceGroupName, mapName, body, context);
     }
 
-    public void exportDependencies(String resourceGroupName, String mapName, ExportDependenciesRequest body) {
-        this.serviceClient().exportDependencies(resourceGroupName, mapName, body);
+    public ExportDependenciesOperationResult exportDependencies(String resourceGroupName, String mapName,
+        ExportDependenciesRequest body) {
+        ExportDependenciesOperationResultInner inner
+            = this.serviceClient().exportDependencies(resourceGroupName, mapName, body);
+        if (inner != null) {
+            return new ExportDependenciesOperationResultImpl(inner, this.manager());
+        } else {
+            return null;
+        }
     }
 
-    public void exportDependencies(String resourceGroupName, String mapName, ExportDependenciesRequest body,
-        Context context) {
-        this.serviceClient().exportDependencies(resourceGroupName, mapName, body, context);
+    public ExportDependenciesOperationResult exportDependencies(String resourceGroupName, String mapName,
+        ExportDependenciesRequest body, Context context) {
+        ExportDependenciesOperationResultInner inner
+            = this.serviceClient().exportDependencies(resourceGroupName, mapName, body, context);
+        if (inner != null) {
+            return new ExportDependenciesOperationResultImpl(inner, this.manager());
+        } else {
+            return null;
+        }
+    }
+
+    public GetDependencyViewForAllMachinesOperationResult getDependencyViewForAllMachines(String resourceGroupName,
+        String mapName, GetDependencyViewForAllMachinesRequest body) {
+        GetDependencyViewForAllMachinesOperationResultInner inner
+            = this.serviceClient().getDependencyViewForAllMachines(resourceGroupName, mapName, body);
+        if (inner != null) {
+            return new GetDependencyViewForAllMachinesOperationResultImpl(inner, this.manager());
+        } else {
+            return null;
+        }
+    }
+
+    public GetDependencyViewForAllMachinesOperationResult getDependencyViewForAllMachines(String resourceGroupName,
+        String mapName, GetDependencyViewForAllMachinesRequest body, Context context) {
+        GetDependencyViewForAllMachinesOperationResultInner inner
+            = this.serviceClient().getDependencyViewForAllMachines(resourceGroupName, mapName, body, context);
+        if (inner != null) {
+            return new GetDependencyViewForAllMachinesOperationResultImpl(inner, this.manager());
+        } else {
+            return null;
+        }
     }
 
     public MapsResource getById(String id) {