+++ sdk/dependencymap/azure-resourcemanager-dependencymap/src/main/java/com/azure/resourcemanager/dependencymap/implementation/MapsResourceImpl.java
@@ -8,9 +8,12 @@ import com.azure.core.management.Region;
 import com.azure.core.management.SystemData;
 import com.azure.core.util.Context;
 import com.azure.resourcemanager.dependencymap.fluent.models.MapsResourceInner;
+import com.azure.resourcemanager.dependencymap.models.ExportDependenciesOperationResult;
 import com.azure.resourcemanager.dependencymap.models.ExportDependenciesRequest;
 import com.azure.resourcemanager.dependencymap.models.GetConnectionsForProcessOnFocusedMachineRequest;
 import com.azure.resourcemanager.dependencymap.models.GetConnectionsWithConnectedMachineForFocusedMachineRequest;
+import com.azure.resourcemanager.dependencymap.models.GetDependencyViewForAllMachinesOperationResult;
+import com.azure.resourcemanager.dependencymap.models.GetDependencyViewForAllMachinesRequest;
 import com.azure.resourcemanager.dependencymap.models.GetDependencyViewForFocusedMachineRequest;
 import com.azure.resourcemanager.dependencymap.models.MapsResource;
 import com.azure.resourcemanager.dependencymap.models.MapsResourceProperties;
@@ -177,12 +180,22 @@ public final class MapsResourceImpl implements MapsResource, MapsResource.Defini
         serviceManager.maps().getConnectionsForProcessOnFocusedMachine(resourceGroupName, mapName, body, context);
     }
 
-    public void exportDependencies(ExportDependenciesRequest body) {
-        serviceManager.maps().exportDependencies(resourceGroupName, mapName, body);
+    public ExportDependenciesOperationResult exportDependencies(ExportDependenciesRequest body) {
+        return serviceManager.maps().exportDependencies(resourceGroupName, mapName, body);
     }
 
-    public void exportDependencies(ExportDependenciesRequest body, Context context) {
-        serviceManager.maps().exportDependencies(resourceGroupName, mapName, body, context);
+    public ExportDependenciesOperationResult exportDependencies(ExportDependenciesRequest body, Context context) {
+        return serviceManager.maps().exportDependencies(resourceGroupName, mapName, body, context);
+    }
+
+    public GetDependencyViewForAllMachinesOperationResult
+        getDependencyViewForAllMachines(GetDependencyViewForAllMachinesRequest body) {
+        return serviceManager.maps().getDependencyViewForAllMachines(resourceGroupName, mapName, body);
+    }
+
+    public GetDependencyViewForAllMachinesOperationResult
+        getDependencyViewForAllMachines(GetDependencyViewForAllMachinesRequest body, Context context) {
+        return serviceManager.maps().getDependencyViewForAllMachines(resourceGroupName, mapName, body, context);
     }
 
     public MapsResourceImpl withRegion(Region location) {