+++ sdk/dependencymap/azure-resourcemanager-dependencymap/src/main/java/com/azure/resourcemanager/dependencymap/models/Maps.java
@@ -193,8 +193,10 @@ public interface Maps {
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the response.
      */
-    void exportDependencies(String resourceGroupName, String mapName, ExportDependenciesRequest body);
+    ExportDependenciesOperationResult exportDependencies(String resourceGroupName, String mapName,
+        ExportDependenciesRequest body);
 
     /**
      * Export dependencies.
@@ -206,8 +208,39 @@ public interface Maps {
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the response.
      */
-    void exportDependencies(String resourceGroupName, String mapName, ExportDependenciesRequest body, Context context);
+    ExportDependenciesOperationResult exportDependencies(String resourceGroupName, String mapName,
+        ExportDependenciesRequest body, Context context);
+
+    /**
+     * Get dependencies for all machines.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param mapName Maps resource name.
+     * @param body The content of the action request.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return dependencies for all machines.
+     */
+    GetDependencyViewForAllMachinesOperationResult getDependencyViewForAllMachines(String resourceGroupName,
+        String mapName, GetDependencyViewForAllMachinesRequest body);
+
+    /**
+     * Get dependencies for all machines.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param mapName Maps resource name.
+     * @param body The content of the action request.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return dependencies for all machines.
+     */
+    GetDependencyViewForAllMachinesOperationResult getDependencyViewForAllMachines(String resourceGroupName,
+        String mapName, GetDependencyViewForAllMachinesRequest body, Context context);
 
     /**
      * Get a MapsResource.